Migrate About Page styles (#1084)
* Migrate styles for About Page - initial commit

* Migrate styles for AssociationMembersSection, extract common styles

* Migrate styles on About page - minor changes

* Fix build

* Migrate styles for ManagementBoardSection

* Migrate styles for SupervisoryBoardSection

* Delete unused image

* Change discord team image, extract AboutWrapper styles

* Migarte styles for PrinciplesThatUniteUs

* Migrate styles for PrincipleCard

* Fix build error

* Fix LinkedIn button
