+++ src/components/about/helpers/supervisoryBoardData.tsx
@@ -9,7 +9,7 @@ export const data: TeamData[] = [
   {
     img: '/img/team-photos/GeorgiMalchev.jpg',
     name: 'Георги Малчев',
-    linkedInProfile: 'https://www.linkedin.com/in/georgimalchev//',
+    linkedInProfile: 'https://www.linkedin.com/in/georgimalchev/',
     description:
       'Преподавател, Съдружник в агенция Xplora.bg. Член на управителния съвет на IAB Bulgaria. Маркетингът е моето призвание. Гордея се, че за почти 6 години от основаването си агенцията за интегриран дигитален маркетинг Xplora, в която съм управляващ съдружник, е сред водещите дигитални агенции в България. Имаме над 90 текущи клиенти - български и мултинационални компании, които са сред топ 3 в сектора си в областта на финансите, IT и технологичните решения, хранително-вкусовата промишленост и търговията.',
   },
