+++ src/components/about/sections/AboutTheTeamSection/AboutTheTeamSection.styled.tsx
@@ -1,0 +1,10 @@
+import { styled } from '@mui/system'
+import { Typography } from '@mui/material'
+
+import theme from 'common/theme'
+
+export const DiscordTeamImage = styled(Typography)(() => ({
+  marginTop: theme.spacing(8),
+  display: 'flex',
+  justifyContent: 'center',
+}))
