+++ src/components/about/sections/PrinciplesThatUniteUsSection/PrinciplesThatUniteUs.tsx
@@ -1,9 +1,13 @@
 import React from 'react'
+
 import { useTranslation } from 'next-i18next'
+
 import { Grid } from '@mui/material'
+
 import PrincipleCard from './PrincipleCard'
-import Heading from 'components/common/Heading'
-import { principlesData } from '../helpers/principlesData'
+import { principlesData } from '../../helpers/principlesData'
+
+import { AboutHeading } from 'components/about/AboutPage.styled'
 
 export default function PrinciplesThatUniteUs() {
   const { t } = useTranslation('about')
@@ -11,9 +15,7 @@ export default function PrinciplesThatUniteUs() {
   return (
     <Grid container spacing={1}>
       <Grid xs={12} item>
-        <Heading variant="h4" component="h2" textAlign="center" fontWeight="500" mt={15} mb={8}>
-          {t('principlesThatUniteUs.title')}
-        </Heading>
+        <AboutHeading variant="h4">{t('principlesThatUniteUs.title')}</AboutHeading>
       </Grid>
       {principlesData.map((principle) => (
         <Grid key={principle.heading} item xs={12} sm={6}>
